refactor: update controllers to use ResponseEntity to return better responses to API

+++ src/main/java/dev/joaov/cadastroNinja/Missoes/MissoesController.java
@@ -1,5 +1,8 @@
 package dev.joaov.cadastroNinja.Missoes;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,28 +19,37 @@ public class MissoesController {
     }
 
     @PostMapping("/create")
-    public MissoesDTO createMission(@RequestBody MissoesDTO missao) {
-        return missoesService.createMission(missao);
+    public ResponseEntity<MissoesDTO> createMission(@RequestBody MissoesDTO missao) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(missoesService.createMission(missao));
     }
 
     @GetMapping("/show")
-    public List<MissoesDTO> showMissions() {
-        return missoesService.showMissions();
+    public ResponseEntity<List<MissoesDTO>> showMissions() {
+        return ResponseEntity.ok(missoesService.showMissions());
     }
 
     @GetMapping("/show/{id}")
-    public MissoesDTO showMissionsById(@PathVariable Long id) {
-        return missoesService.showMissionsById(id);
+    public ResponseEntity<Object> showMissionsById(@PathVariable Long id) {
+        if (missoesService.showMissionsById(id) == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Mission with id " + id + " not found.");
+        }
+        return ResponseEntity.ok(missoesService.showMissionsById(id));
     }
 
     @PutMapping("/update/{id}")
-    public MissoesDTO updateMission(@PathVariable Long id, @RequestBody MissoesDTO missao) {
-        return missoesService.updateMission(id, missao);
+    public ResponseEntity<Object> updateMission(@PathVariable Long id, @RequestBody MissoesDTO missao) {
+        if (missoesService.updateMission(id, missao) == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Mission with id " + id + " not found.");
+        }
+        return ResponseEntity.ok(missoesService.updateMission(id, missao));
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteMission(@PathVariable Long id) {
+    public ResponseEntity<String> deleteMission(@PathVariable Long id) {
+        if (missoesService.showMissionsById(id) == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Mission with id " + id + " not found.");
+        }
         missoesService.deleteMission(id);
+        return ResponseEntity.ok("Mission with id " + id + " deleted successfully.");
     }
-
 }
